+++ data-browser/src/views/RedirectPage.tsx
@@ -5,7 +5,7 @@ import { useHistory } from 'react-router-dom';
 
 import { ContainerNarrow } from '../components/Containers';
 import { ValueForm } from '../components/forms/ValueForm';
-import { editURL, openURL } from '../helpers/navigation';
+import { openURL } from '../helpers/navigation';
 import AllProps from '../components/AllProps';
 import { useSettings } from '../helpers/AppSettings';
 import { Button } from '../components/Button';